+++ source/common/config/utility.cc
@@ -3,6 +3,7 @@
 #include <unordered_set>
 
 #include "envoy/config/metrics/v2/stats.pb.h"
+#include "envoy/stats/scope.h"
 
 #include "common/common/assert.h"
 #include "common/common/fmt.h"
@@ -157,6 +158,12 @@ std::chrono::milliseconds Utility::apiConfigSourceRefreshDelay(
 
   return std::chrono::milliseconds(
       DurationUtil::durationToMilliseconds(api_config_source.refresh_delay()));
+}
+
+std::chrono::milliseconds Utility::apiConfigSourceRequestTimeout(
+    const envoy::api::v2::core::ApiConfigSource& api_config_source) {
+  return std::chrono::milliseconds(
+      PROTOBUF_GET_MS_OR_DEFAULT(api_config_source, request_timeout, 1000));
 }
 
 void Utility::translateEdsConfig(const Json::Object& json_config,