+++ source/extensions/filters/http/dynamo/dynamo_filter.h
@@ -5,7 +5,7 @@
 
 #include "envoy/http/filter.h"
 #include "envoy/runtime/runtime.h"
-#include "envoy/stats/stats.h"
+#include "envoy/stats/scope.h"
 
 #include "common/json/json_loader.h"