+++ source/server/hot_restart_impl.h
@@ -10,6 +10,7 @@
 
 #include "envoy/server/hot_restart.h"
 #include "envoy/server/options.h"
+#include "envoy/stats/stats_options.h"
 
 #include "common/common/assert.h"
 #include "common/common/block_memory_hash_set.h"