+++ source/server/overload_manager_impl.cc
@@ -1,5 +1,7 @@
 #include "server/overload_manager_impl.h"
 
+#include "envoy/stats/scope.h"
+
 #include "common/common/fmt.h"
 #include "common/config/utility.h"
 #include "common/protobuf/utility.h"
@@ -128,6 +130,9 @@ OverloadManagerImpl::OverloadManagerImpl(
 void OverloadManagerImpl::start() {
   ASSERT(!started_);
   started_ = true;
+  if (resources_.empty()) {
+    return;
+  }
   timer_ = dispatcher_.createTimer([this]() -> void {
     for (auto& resource : resources_) {
       resource.second.update();