+++ test/common/config/utility_test.cc
@@ -1,6 +1,5 @@
 #include "envoy/api/v2/eds.pb.h"
 #include "envoy/common/exception.h"
-#include "envoy/stats/stats.h"
 
 #include "common/common/fmt.h"
 #include "common/config/cds_json.h"
@@ -57,6 +56,18 @@ TEST(UtilityTest, ApiConfigSourceRefreshDelay) {
   api_config_source.mutable_refresh_delay()->CopyFrom(
       Protobuf::util::TimeUtil::MillisecondsToDuration(1234));
   EXPECT_EQ(1234, Utility::apiConfigSourceRefreshDelay(api_config_source).count());
+}
+
+TEST(UtilityTest, ApiConfigSourceDefaultRequestTimeout) {
+  envoy::api::v2::core::ApiConfigSource api_config_source;
+  EXPECT_EQ(1000, Utility::apiConfigSourceRequestTimeout(api_config_source).count());
+}
+
+TEST(UtilityTest, ApiConfigSourceRequestTimeout) {
+  envoy::api::v2::core::ApiConfigSource api_config_source;
+  api_config_source.mutable_request_timeout()->CopyFrom(
+      Protobuf::util::TimeUtil::MillisecondsToDuration(1234));
+  EXPECT_EQ(1234, Utility::apiConfigSourceRequestTimeout(api_config_source).count());
 }
 
 TEST(UtilityTest, TranslateApiConfigSource) {