+++ test/common/filesystem/filesystem_impl_test.cc
@@ -1,8 +1,6 @@
 #include <chrono>
 #include <string>
 
-#include "envoy/stats/stats.h"
-
 #include "common/api/os_sys_calls_impl.h"
 #include "common/common/lock_guard.h"
 #include "common/common/thread.h"