+++ test/common/network/listener_impl_test.cc
@@ -1,5 +1,3 @@
-#include "envoy/stats/stats.h"
-
 #include "common/network/address_impl.h"
 #include "common/network/listener_impl.h"
 #include "common/network/utility.h"